+++ examples/neodriver_test.rs
@@ -50,20 +50,9 @@ fn main() -> ! {
 
     let id = neo_driver.hardware_id().expect("Failed to get hardware id");
     rprintln!("Hardware ID {:?}", id);
-    rprintln!(
-        "Product info {:#?}",
-        neo_driver
-            .product_info()
-            .expect("failed to get product info")
-    );
 
     let mut colors: [rgb::Grb<u8>; N_LEDS] =
         core::array::from_fn(|i| color_wheel((i as u8).wrapping_mul(255 / N_LEDS as u8)));
-
-    neo_driver
-        .set_neopixel_colors(&colors)
-        .and_then(|_| neo_driver.sync_neopixel())
-        .expect("Failed to set neopixel colors");
 
     loop {
         colors.rotate_left(1);