+++ src/devices/generic_device.rs
@@ -1,7 +1,7 @@
 use super::SeesawDeviceInit;
 use crate::{
     modules::{status::StatusModule, HardwareId},
-    seesaw_device, Driver,
+    seesaw_device, Driver, SeesawError,
 };
 
 seesaw_device! {
@@ -9,12 +9,11 @@ seesaw_device! {
   name: GenericDevice,
   hardware_id: HardwareId::SAMD09,
   product_id: 0,
-  default_addr: 0x49,
-  modules: []
+  default_addr: 0x49
 }
 
 impl<D: Driver> SeesawDeviceInit<D> for GenericDevice<D> {
-    fn init(mut self) -> Result<Self, Self::Error> {
+    fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
         self.reset().map(|_| self)
     }
 }