+++ src/devices/macros.rs
@@ -5,14 +5,7 @@ macro_rules! seesaw_device {
         name: $name:ident,
         hardware_id: $hardware_id:expr,
         product_id: $product_id:expr,
-        default_addr: $default_addr:expr,
-        modules: [
-            $($module_name:ident$(<$($module_param_name:ident =$module_param:ty),*>)? $({
-                $($const_name:ident: $const_value:expr),*
-            })?),*
-            $(,)?
-        ]
-         $(,)?
+        default_addr: $default_addr:expr
     ) => {
         #[doc=core::concat!("[Adafruit Product Page](https://www.adafruit.com/product/", core::stringify!($product_id),")")]
         #[doc=core::concat!("")]
@@ -34,7 +27,6 @@ macro_rules! seesaw_device {
 
         impl<D: $crate::Driver> $crate::devices::SeesawDevice for $name<D> {
             type Driver = D;
-            type Error = $crate::SeesawError<D::Error>;
             const DEFAULT_ADDR: u8 = $default_addr;
             const HARDWARE_ID: $crate::modules::HardwareId = $hardware_id;
             const PRODUCT_ID: u16 = $product_id;
@@ -55,57 +47,5 @@ macro_rules! seesaw_device {
                 Self(Self::DEFAULT_ADDR, driver)
             }
         }
-
-        $(
-            impl_device_module! {
-                $name, $module_name$(<$($module_param_name = $module_param),*>)? $({
-                    $($const_name: $const_value),*
-                })*
-            }
-        )*
-    };
-}
-
-#[doc(hidden)]
-#[macro_export(local_inner_macros)]
-macro_rules! impl_device_module {
-    ($device:ident, AdcModule $({})?) => {
-        impl<D: $crate::Driver> $crate::modules::adc::AdcModule<D> for $device<D> {}
-    };
-    ($device:ident, EncoderModule {
-        num_encoders: $num_encoders:expr,
-        encoder_btn_pins: $button_pins:expr
-    }) => {
-        impl<D: $crate::Driver> $crate::modules::encoder::EncoderModule<D, $num_encoders>
-            for $device<D>
-        {
-            const ENCODER_BTN_PINS: [u8; $num_encoders] = $button_pins;
-        }
-    };
-    ($device:ident, GpioModule $({})?) => {
-        impl<D: $crate::Driver> $crate::modules::gpio::GpioModule<D> for $device<D> {}
-    };
-    ($device:ident, KeypadModule { num_cols: $num_cols:expr, num_rows: $num_rows:expr }) => {
-        impl<D: $crate::Driver> $crate::modules::keypad::KeypadModule<D> for $device<D> {
-            const NUM_COLS: u8 = $num_cols;
-            const NUM_ROWS: u8 = $num_rows;
-        }
-    };
-    ($device:ident, NeopixelModule<color_type = $color_type:ty> { num_leds: $num_leds:expr, pin: $pin:expr }) => {
-        impl<D: $crate::Driver> $crate::modules::neopixel::NeopixelModule<D> for $device<D> {
-            type Color = $color_type;
-
-            const N_LEDS: usize = $num_leds;
-            const PIN: u8 = $pin;
-        }
-    };
-    ($device:ident, StatusModule $({})?) => {
-        impl<D: $crate::Driver> $crate::modules::StatusModule<D> for $device<D> {}
-    };
-    ($device:ident, TimerModule $({})?) => {
-        impl<D: $crate::Driver> $crate::modules::timer::TimerModule<D> for $device<D> {}
-    };
-    ($device:ident, QuadEncoderModule $({})?) => {
-        impl<D: $crate::Driver> $crate::modules::quad_encoder::QuadEncoderModule<D> for $device<D> {}
     };
 }