+++ src/devices/neokey_1x4.rs
@@ -14,17 +14,21 @@ seesaw_device! {
   name: NeoKey1x4,
   hardware_id: HardwareId::SAMD09,
   product_id: 4980,
-  default_addr: 0x30,
-  modules: [
-      GpioModule,
-      NeopixelModule<color_type = NeoKey1x4Color> { num_leds: 4, pin: 3 },
-  ]
+  default_addr: 0x30
 }
 
 pub type NeoKey1x4Color = rgb::Grb<u8>;
 
+impl<D: Driver> GpioModule<D> for NeoKey1x4<D> {}
+impl<D: Driver> NeopixelModule<D> for NeoKey1x4<D> {
+    type Color = NeoKey1x4Color;
+
+    const N_LEDS: usize = 4;
+    const PIN: u8 = 3;
+}
+
 impl<D: Driver> SeesawDeviceInit<D> for NeoKey1x4<D> {
-    fn init(mut self) -> Result<Self, Self::Error> {
+    fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
         self.reset_and_verify_seesaw()
             .and_then(|_| self.enable_neopixel())
             .and_then(|_| self.enable_button_pins())