+++ src/devices/neoslider.rs
@@ -1,6 +1,9 @@
 use super::SeesawDeviceInit;
 use crate::{
-    modules::{adc::AdcModule, neopixel::NeopixelModule, status::StatusModule, HardwareId},
+    modules::{
+        adc::AdcModule, gpio::GpioModule, neopixel::NeopixelModule, status::StatusModule,
+        HardwareId,
+    },
     seesaw_device, Driver, SeesawError,
 };
 
@@ -9,18 +12,22 @@ seesaw_device!(
   name: NeoSlider,
   hardware_id: HardwareId::ATTINY817,
   product_id: 5295,
-  default_addr: 0x30,
-  modules: [
-      AdcModule,
-      GpioModule,
-      NeopixelModule<color_type = NeoSliderColor> {  num_leds: 4, pin: 14 },
-  ]
+  default_addr: 0x30
 );
 
 pub type NeoSliderColor = rgb::Grb<u8>;
 
+impl<D: Driver> AdcModule<D> for NeoSlider<D> {}
+impl<D: Driver> GpioModule<D> for NeoSlider<D> {}
+impl<D: Driver> NeopixelModule<D> for NeoSlider<D> {
+    type Color = NeoSliderColor;
+
+    const N_LEDS: usize = 4;
+    const PIN: u8 = 14;
+}
+
 impl<D: Driver> SeesawDeviceInit<D> for NeoSlider<D> {
-    fn init(mut self) -> Result<Self, Self::Error> {
+    fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
         self.reset_and_verify_seesaw()
             .and_then(|_| self.enable_neopixel())
             .map(|_| self)