+++ src/devices/neotrellis.rs
@@ -8,17 +8,25 @@ seesaw_device! {
     name: NeoTrellis,
     hardware_id: HardwareId::SAMD09,
     product_id: 3954,
-    default_addr: 0x2E,
-    modules: [
-        NeopixelModule<color_type = NeoTrellisColor> { num_leds: 16, pin: 3 },
-        KeypadModule { num_cols: 4, num_rows: 4 },
-    ]
+    default_addr: 0x2E
 }
 
 pub type NeoTrellisColor = rgb::Grb<u8>;
 
+impl<D: Driver> KeypadModule<D> for NeoTrellis<D> {
+    const NUM_COLS: u8 = 4;
+    const NUM_ROWS: u8 = 4;
+}
+
+impl<D: Driver> NeopixelModule<D> for NeoTrellis<D> {
+    type Color = NeoTrellisColor;
+
+    const N_LEDS: usize = 16;
+    const PIN: u8 = 3;
+}
+
 impl<D: Driver> SeesawDeviceInit<D> for NeoTrellis<D> {
-    fn init(mut self) -> Result<Self, Self::Error> {
+    fn init(mut self) -> Result<Self, SeesawError<D::Error>> {
         self.reset_and_verify_seesaw()
             .and_then(|_| self.enable_neopixel())
             .map(|_| self)
@@ -26,14 +34,6 @@ impl<D: Driver> SeesawDeviceInit<D> for NeoTrellis<D> {
 }
 
 impl<D: Driver> NeoTrellis<D> {
-    pub const fn num_cols(&self) -> u8 {
-        Self::NUM_COLS
-    }
-
-    pub const fn num_rows(&self) -> u8 {
-        Self::NUM_ROWS
-    }
-
     pub fn set_xy_neopixel_color(
         &mut self,
         x: u8,