+++ src/modules/neopixel.rs
@@ -1,7 +1,6 @@
-use rgb::ComponentSlice;
-
 use super::{Modules, Reg};
 use crate::{devices::SeesawDevice, driver::Driver, DriverExt, SeesawError};
+use rgb::ComponentSlice;
 
 /// WO - 8 bits
 /// This register sets the pin number (PORTA) that is used for the NeoPixel
@@ -26,7 +25,13 @@ const SHOW: &Reg = &[Modules::Neopixel.into_u8(), 0x05];
 
 pub trait NeopixelModule<D: Driver>: SeesawDevice<Driver = D> {
     /// The size of the color type in bytes
-    const C_SIZE: usize = core::mem::size_of::<Self::Color>();
+    const C_SIZE: usize = {
+        match core::mem::size_of::<Self::Color>() {
+            3 => 3,
+            4 => 4,
+            _ => panic!("Invalid color size"),
+        }
+    };
     /// The number of neopixels on or connected to the device
     const N_LEDS: usize = 1;
     /// The output pin of the neopixel signal
@@ -95,24 +100,42 @@ pub trait NeopixelModule<D: Driver>: SeesawDevice<Driver = D> {
     }
 
     /// Set the color of all neopixels
-    /// TODO: there is room for optimization here (https://github.com/alexeden/adafruit-seesaw/pull/12)
+    ///
+    /// Minimizes the number of transactions performed by chunking the `colors`
+    /// array into the largest (max 30 byte) buffers possible
+    ///
+    /// Note that if `C_SIZE` is _not_ 3 or 4 bytes, the chunking optimization
+    /// is effectively skipped
     fn set_neopixel_colors(
         &mut self,
         colors: &[Self::Color; Self::N_LEDS],
     ) -> Result<(), SeesawError<D::Error>>
     where
-        [(); 2 + Self::C_SIZE]: Sized,
+        [(); 2 + color_bytes_per_write(Self::C_SIZE)]: Sized,
     {
+        let mut buf = [0; 2 + color_bytes_per_write(Self::C_SIZE)];
         let addr = self.addr();
-        let mut buf = [0; 2 + Self::C_SIZE];
-        for (n, color) in colors.iter().enumerate() {
-            buf[..2].copy_from_slice(&u16::to_be_bytes((Self::C_SIZE * n) as u16));
-            buf[2..].copy_from_slice(color.as_slice());
-            self.driver()
-                .register_write(addr, SET_BUF, &buf)
-                .map_err(SeesawError::I2c)?;
-        }
-        Ok(())
+
+        colors
+            .chunks(max_colors_per_write(Self::C_SIZE))
+            .enumerate()
+            .try_for_each(|(i, chunk)| {
+                let offset = u16::to_be_bytes(
+                    (Self::C_SIZE * i * max_colors_per_write(Self::C_SIZE)) as u16,
+                );
+                buf[..2].copy_from_slice(&offset);
+                chunk.iter().enumerate().for_each(|(j, c)| {
+                    let start = 2 + (j * Self::C_SIZE);
+                    buf[start..start + Self::C_SIZE].copy_from_slice(c.as_slice());
+                });
+
+                self.driver().register_write(
+                    addr,
+                    SET_BUF,
+                    &buf[0..2 + (Self::C_SIZE * chunk.len())],
+                )
+            })
+            .map_err(SeesawError::I2c)
     }
 
     fn sync_neopixel(&mut self) -> Result<(), SeesawError<D::Error>> {
@@ -123,6 +146,22 @@ pub trait NeopixelModule<D: Driver>: SeesawDevice<Driver = D> {
             .map(|_| self.driver().delay_us(125))
             .map_err(SeesawError::I2c)
     }
+}
+
+/// Get the maximum number of colors that can be written in a single write
+/// operation as a function of the number of bytes per color
+pub const fn max_colors_per_write(c_size: usize) -> usize {
+    match c_size {
+        3 => 9, // 27
+        4 => 7, // 28
+        _ => 1, // effectively skips the optimization
+    }
+}
+
+/// Get the number of bytes dedicated to writing colors in a single write
+/// operation as a function of the number of bytes per color
+pub const fn color_bytes_per_write(c_size: usize) -> usize {
+    c_size * max_colors_per_write(c_size)
 }
 
 /// NeopixelModule: The Neopixel protocol speed