+++ src/modules/timer.rs
@@ -10,9 +10,13 @@ const PWM_VAL: &Reg = &[Modules::Timer.into_u8(), 0x01];
 /// The module base register address for the PWM module is 0x08.
 /// PWM outputs are available on pins PA04, PA05, PA06, and PA07.
 pub trait TimerModule<D: Driver>: SeesawDevice<Driver = D> {
+    /// Write a PWM value to a PWM-enabled pin
+    ///
+    /// On the SAMD09 breakout, the pin corresponds to the number on the
+    /// silkscreen. On the default seesaw firmware on the SAMD09 breakout,
+    /// pins 5, 6, and 7 are PWM enabled.
     fn analog_write(&mut self, pin: u8, value: u8) -> Result<(), SeesawError<D::Error>> {
         let mapped_pin = match Self::HARDWARE_ID {
-            HardwareId::ATTINY817 => pin,
             HardwareId::SAMD09 => match pin {
                 4 => 0,
                 5 => 1,
@@ -20,6 +24,7 @@ pub trait TimerModule<D: Driver>: SeesawDevice<Driver = D> {
                 7 => 3,
                 _ => 0,
             },
+            _ => pin,
         };
 
         let addr = self.addr();